// server/db.d.ts
/**
 * This file was generated by kysely-codegen.
 * Please do not edit it manually.
 */

import type { Cookie } from "express-session";
import type { ColumnType } from "kysely";

export type Generated<T> =
  T extends ColumnType<infer S, infer I, infer U>
    ? ColumnType<S, I | undefined, U>
    : ColumnType<T, T | undefined, T>;

export type Json = JsonValue;

export type JsonArray = JsonValue[];

export type JsonObject = {
  [x: string]: JsonValue | undefined;
};

export type JsonPrimitive = boolean | number | string | null;

export type JsonValue = JsonArray | JsonObject | JsonPrimitive;

export type Timestamp = ColumnType<Date, Date | string, Date | string>;

export interface Cart {
  created_at: Generated<Timestamp>;
  id: Generated<string>;
  session_id: string | null;
  updated_at: Generated<Timestamp>;
  user_id: string | null;
}

export interface CartItems {
  cart_id: string | null;
  created_at: Generated<Timestamp>;
  id: Generated<string>;
  price: number | null;
  product_id: string | null;
  quantity: number | null;
  total_price: Generated<number | null>;

  updated_at: Generated<Timestamp>;
}

export interface Categories {
  created_at: Generated<Timestamp>;
  description: string | null;
  details: Json | null;
  id: Generated<string>;
  images: string[] | null;
  name: string;
}

export interface OrderItems {
  created_at: Generated<Timestamp>;
  id: Generated<string>;
  order_id: string | null;
  price: number | null;
  product_id: string | null;
  quantity: number | null;
  updated_at: Generated<Timestamp>;
}

export interface Orders {
  address_id: string | null;
  created_at: Generated<Timestamp>;
  id: Generated<string>;
  status: Generated<string | null>;
  updated_at: Generated<Timestamp>;
  user_id: string | null;
  value: number;
}

export interface ProductCategory {
  amount: number | null;
  category_id: string;
  product_id: string;
}

export interface Products {
  created_at: Generated<Timestamp>;
  description: string;
  details: Json | null;
  id: Generated<string>;
  images: {
    public_id: string;

    width: number;
    height: number;
    format: string;
    created_at: Generated<Timestamp>;
    bytes: number;
    url: string;
    secure_url: string;
  } | null;
  name: string;
  price: number;
  user_id: string | null;
}

export interface Addresses {
  address_one: string;
  address_two: string;
  city: string;
  country: string;
  created_at: Generated<Timestamp>;
  email: string;
  id: Generated<string>;
  full_name: string;
  label: string;
  phonenumber: number;
  province: string;
  user_id: string | null;
  zipcode: string;
}

export interface Users {
  created_at: Generated<Timestamp>;
  email: string;
  id: Generated<string>;
  default_address_id: Generated<string> | null;
  name: string;
  password: string;
  admin: boolean;
  cart_id: string;
}

export interface Frontpage {
  categories: Categories[] | null;
  id: Generated<string>;
  current: boolean | null;
  name: string;
  products: Products[] | null;
}

export interface Session {
  sid: string;
  sess: {
    cookie?: Cookie;
    messages?: string[];
    passport?: {
      user?: string;
    };
    cart_id?: string;
  };
  expire: Generated<Timestamp>;
}

export interface DB {
  addresses: Addresses;
  cart: Cart;
  cart_items: CartItems;
  categories: Categories;
  order_items: OrderItems;
  orders: Orders;
  product_category: ProductCategory;
  products: Products;
  frontpage: Frontpage;
  session: Session;
  users: Users;
}